Feat: #66 시설등록 강사 추가

+++ backend/src/main/java/com/ohammer/apartner/domain/facility/entity/FacilityTimeSlot.java
@@ -42,6 +42,6 @@ public class FacilityTimeSlot extends BaseEntity {
     private LocalTime endTime; // 10:30
 
     @Column(name = "max_capacity", nullable = false)
-    private Integer maxCapacity; // 슬롯별 최대 예약인원
+    private Long maxCapacity; // 슬롯별 최대 예약인원
 
 }
